work in progress for simulation and other things

+++ java-algo/src/main/java/com/c1games/terminal/myalgo/ExpectedDefense.java
@@ -1,6 +1,5 @@
 package com.c1games.terminal.myalgo;
 
-import com.c1games.terminal.algo.Config;
 import com.c1games.terminal.algo.Coords;
 import com.c1games.terminal.algo.map.GameState;
 import com.c1games.terminal.algo.units.UnitType;
@@ -13,7 +12,7 @@ public class ExpectedDefense {
   /**
    * The total structure health that needs to be killed (or expected.. depending on use case)
    */
-  int structureHealth;
+  float structureHealth;
   /**
    * The expected damage that each unit type would receive if it followed the related path for this defense
    */
@@ -21,7 +20,7 @@ public class ExpectedDefense {
   double expectedDemolisherDamage;
   double expectedIntercepterDamage;
 
-  public ExpectedDefense(GameState move, Coords[] path, int structureHealth, int expectedDamage) {
+  public ExpectedDefense(GameState move, Coords[] path, float structureHealth, float expectedDamage) {
     this.path = path;
     this.structureHealth = structureHealth;
     this.expectedScoutDamage = expectedDamage * move.config.unitInformation.get(UnitType.Scout.ordinal()).speed.orElse(1);
